+++ apm-sniffer/optional-plugins/optional-spring-plugins/optional-spring-cloud/gateway-2.1.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/spring/cloud/gateway/v21x/define/AbstractGateway210EnhancePluginDefine.java
@@ -24,6 +24,7 @@ public abstract class AbstractGateway210EnhancePluginDefine extends ClassInstanc
     @Override
     protected String[] witnessClasses() {
         return new String[] {
+            "org.springframework.cloud.gateway.filter.LoadBalancerClientFilter",
             "org.springframework.cloud.gateway.config.GatewayEnvironmentPostProcessor"
         };
     }