+++ apm-sniffer/optional-plugins/optional-spring-plugins/optional-spring-cloud/gateway-2.1.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/spring/cloud/gateway/v21x/define/ServerWebExchangeInstrumentation.java
@@ -23,6 +23,7 @@ public class ServerWebExchangeInstrumentation extends org.apache.skywalking.apm.
     @Override
     protected String[] witnessClasses() {
         return new String[] {
+            "org.springframework.cloud.gateway.filter.LoadBalancerClientFilter",
             "org.springframework.cloud.gateway.config.GatewayEnvironmentPostProcessor"
         };
     }